[Refactor] 녹음 제출 API와 폴링 API를 하나의 컨트롤러로 통합 - 더이상 세션 단위 녹음 제출이 아닌 개별 녹음 제출이므로, InterviewSessionController 대신 RecordingController라는 이름이 더 적합하다고 판단하여 변경. - API 엔드포인트에서 불필요한 sessionId 삭제
